Changing some halfs to floats in BRDF related code (#3647)
* changing some halfs to floats in BRDF related code, to avoid artifacts and circular banding and making SafeNormalization method type specific

* excluding half3 SafeNormalize for PSSL

* reverting some changes and updating according to reviewer suggestions

* renaming halfDir as it's float now

* reviewer update and changing other Lit shader passes to have same changes as it reproduces as well

* porting changes to UniversalRP from PR to LW

* Changing NoH and d vars to float to avoid NaN's on mobile

* Update CHANGELOG.md

* Update CHANGELOG.md

+++ com.unity.render-pipelines.universal/Shaders/SimpleLitForwardPass.hlsl
@@ -21,12 +21,12 @@ struct Varyings
     float3 posWS                    : TEXCOORD2;    // xyz: posWS
 
 #ifdef _NORMALMAP
-    half4 normal                    : TEXCOORD3;    // xyz: normal, w: viewDir.x
-    half4 tangent                   : TEXCOORD4;    // xyz: tangent, w: viewDir.y
-    half4 bitangent                  : TEXCOORD5;    // xyz: bitangent, w: viewDir.z
+    float4 normal                   : TEXCOORD3;    // xyz: normal, w: viewDir.x
+    float4 tangent                  : TEXCOORD4;    // xyz: tangent, w: viewDir.y
+    float4 bitangent                : TEXCOORD5;    // xyz: bitangent, w: viewDir.z
 #else
-    half3  normal                   : TEXCOORD3;
-    half3 viewDir                   : TEXCOORD4;
+    float3  normal                  : TEXCOORD3;
+    float3 viewDir                  : TEXCOORD4;
 #endif
 
     half4 fogFactorAndVertexLight   : TEXCOORD6; // x: fogFactor, yzw: vertex light
